requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.30.0
bidict==0.23.1
click==8.2.1
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
idna==3.10
numpy==2.3.2
pandas==2.3.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-engineio==4.12.2
python-multipart==0.0.20
python-socketio==5.13.0
pytz==2025.2
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uvicorn==0.35.0
wsproto==1.2.0
